+++ _dist/lib/es6kadoo.mjs
@@ -13,6 +13,7 @@
 /* global define */
 /* eslint no-shadow: ['error', { 'allow': ['root'] }] */
 /* eslint strict: ["error", "function"] */
+const $__ES6GLOB = {};
 (function(root, factory) {
   'use strict';
 
@@ -31,7 +32,7 @@
     /* eslint-disable-next-line no-param-reassign */
     root.ES6Kadoo = factory(root);
   }
-}(this, (root) => {
+}($__ES6GLOB, (root) => {
   'use strict';
 
   /* ***************************************************************************
@@ -342,3 +343,6 @@
   // Returns the library name:
   return $__TREE.src.prototypal;
 }));
+
+// -- Export
+export default $__ES6GLOB.ES6Kadoo