+++ src/basicplus.js
@@ -17,6 +17,7 @@
  *
  * Public Static Methods:
  *  . noConflict                  returns a reference to this ES6Kadoo object,
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
@@ -80,13 +81,14 @@ const ES6Kadoo = {
 
   // Useful to retrieve the library name and version when it is
   // embedded in another library as an object:
-  library: { name: '{{lib:name}}', version: '{{lib:version}}' },
+  _library: { name: '{{lib:name}}', version: '{{lib:version}}' },
 
 
   // -- Private Static Methods ---------------------------------------------
 
   /**
    * Returns the internal objects for testing purpose.
+   * (must not be deleted)
    *
    * @method ()
    * @private
@@ -103,6 +105,7 @@ const ES6Kadoo = {
 
   /**
    * Returns a reference to this ES6Kadoo object.
+   * (must not be deleted)
    *
    * Nota:
    * Running ES6Kadoo in noConflic mode, returns the ES6Kadoo variable to
@@ -114,15 +117,29 @@ const ES6Kadoo = {
    * @returns {Object}      returns the ES6Kadoo object,
    * @since 0.0.0
    */
-  /* istanbul ignore next */
   noConflict() {
     /* eslint-disable-next-line no-param-reassign */
     root.ES6Kadoo = previousES6Kadoo;
     return this;
   },
+
+  /**
+   * Returns the library name and version.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {Object}      returns the library name and version,
+   * @since 0.0.0
+   */
+  whoami() {
+    return this._library;
+  },
 };
 
-// Attaches a constant to ES6Kadoo that provides the version of the lib.
+// Attaches constants to ES6Kadoo that provide name and version of the lib.
+ES6Kadoo.NAME = '{{lib:name}}';
 ES6Kadoo.VERSION = '{{lib:version}}';