+++ src/functional-shared.js
@@ -23,6 +23,7 @@
  *
  *
  * Public Methods:
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
@@ -96,7 +97,7 @@ function extend(object, m) {
  */
 const ES6Kadoo = function(name) {
   const obj = {};
-  obj.library = {
+  obj._library = {
     name: '{{lib:name}}',
     version: '{{lib:version}}',
   };
@@ -105,7 +106,8 @@ const ES6Kadoo = function(name) {
   return obj;
 };
 
-// Attaches a constant to ES6Kadoo that provides the version of the lib.
+// Attaches constants to ES6Kadoo that provide name and version of the lib.
+ES6Kadoo.NAME = '{{lib:name}}';
 ES6Kadoo.VERSION = '{{lib:version}}';
 
 
@@ -113,6 +115,7 @@ ES6Kadoo.VERSION = '{{lib:version}}';
 
 /**
  * Returns the internal objects for testing purpose.
+ * (must not be deleted)
  *
  * @method ()
  * @private
@@ -129,6 +132,7 @@ ES6Kadoo._setTestMode = function() {
 
 /**
  * Returns a reference to this ES6Kadoo object.
+ * (must not be deleted)
  *
  * Nota:
  * Running ES6Kadoo in noConflic mode, returns the ES6Kadoo variable to
@@ -140,7 +144,6 @@ ES6Kadoo._setTestMode = function() {
  * @returns {Object}        returns the ES6Kadoo object,
  * @since 0.0.0
  */
-/* istanbul ignore next */
 ES6Kadoo.noConflict = function() {
   /* eslint-disable-next-line no-param-reassign */
   root.ES6Kadoo = previousES6Kadoo;
@@ -151,6 +154,20 @@ ES6Kadoo.noConflict = function() {
 // -- Public Methods -------------------------------------------------------
 
 methods = {
+
+  /**
+   * Returns the library name and version.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {Object}      returns the library name and version,
+   * @since 0.0.0
+   */
+  whoami() {
+    return this._library;
+  },
 
   /**
    * Returns a string.