+++ src/functional.js
@@ -23,6 +23,7 @@
  *
  *
  * Public Methods:
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
@@ -70,7 +71,7 @@ const previousES6Kadoo = root.ES6Kadoo
  * @since 0.0.0
  */
 const ES6Kadoo = function(name) {
-  obj.library = {
+  obj._library = {
     name: '{{lib:name}}',
     version: '{{lib:version}}',
   };
@@ -78,7 +79,8 @@ const ES6Kadoo = function(name) {
   return obj;
 };
 
-// Attaches a constant to ES6Kadoo that provides the version of the lib.
+// Attaches constants to ES6Kadoo that provide name and version of the lib.
+ES6Kadoo.NAME = '{{lib:name}}';
 ES6Kadoo.VERSION = '{{lib:version}}';
 
 
@@ -86,6 +88,7 @@ ES6Kadoo.VERSION = '{{lib:version}}';
 
 /**
  * Returns the internal objects for testing purpose.
+ * (must not be deleted)
  *
  * @method ()
  * @private
@@ -102,6 +105,7 @@ ES6Kadoo._setTestMode = function() {
 
 /**
  * Returns a reference to this ES6Kadoo object.
+ * (must not be deleted)
  *
  * Nota:
  * Running ES6Kadoo in noConflic mode, returns the ES6Kadoo variable to
@@ -113,7 +117,6 @@ ES6Kadoo._setTestMode = function() {
  * @returns {Object}        returns the ES6Kadoo object,
  * @since 0.0.0
  */
-/* istanbul ignore next */
 ES6Kadoo.noConflict = function() {
   /* eslint-disable-next-line no-param-reassign */
   root.ES6Kadoo = previousES6Kadoo;
@@ -122,6 +125,20 @@ ES6Kadoo.noConflict = function() {
 
 
 // -- Public Methods -------------------------------------------------------
+
+/**
+ * Returns the library name and version.
+ * (must not be deleted)
+ *
+ * @method ()
+ * @public
+ * @param {}            -,
+ * @returns {Object}    returns the library name and version,
+ * @since 0.0.0
+ */
+obj.whoami = function() {
+  return obj._library;
+};
 
 /**
  * Returns a string.