+++ src/pseudoclassical-auto.js
@@ -24,6 +24,7 @@
  *
  *
  * Public Methods:
+ *  . whoami                      returns the library name and version,
  *  . getString                   returns a string,
  *  . getArray                    returns an array,
  *
@@ -71,7 +72,7 @@ const previousES6Kadoo = root.ES6Kadoo
  */
 const ES6Kadoo = function(name) {
   if (this instanceof ES6Kadoo) {
-    this.library = {
+    this._library = {
       name: '{{lib:name}}',
       version: '{{lib:version}}',
     };
@@ -82,7 +83,8 @@ const ES6Kadoo = function(name) {
   return null;
 };
 
-// Attaches a constant to ES6Kadoo that provides the version of the lib.
+// Attaches constants to ES6Kadoo that provide name and version of the lib.
+ES6Kadoo.NAME = '{{lib:name}}';
 ES6Kadoo.VERSION = '{{lib:version}}';
 
 
@@ -90,6 +92,7 @@ ES6Kadoo.VERSION = '{{lib:version}}';
 
 /**
  * Returns the internal objects for testing purpose.
+ * (must not be deleted)
  *
  * @method ()
  * @private
@@ -106,6 +109,7 @@ ES6Kadoo._setTestMode = function() {
 
 /**
  * Returns a reference to this ES6Kadoo object.
+ * (must not be deleted)
  *
  * Nota:
  * Running ES6Kadoo in noConflic mode, returns the ES6Kadoo variable to
@@ -117,7 +121,6 @@ ES6Kadoo._setTestMode = function() {
  * @returns {Object}        returns the ES6Kadoo object,
  * @since 0.0.0
  */
-/* istanbul ignore next */
 ES6Kadoo.noConflict = function() {
   /* eslint-disable-next-line no-param-reassign */
   root.ES6Kadoo = previousES6Kadoo;
@@ -128,6 +131,20 @@ ES6Kadoo.noConflict = function() {
 // -- Public Methods -------------------------------------------------------
 
 ES6Kadoo.prototype = {
+
+  /**
+   * Returns the library name and version.
+   * (must not be deleted)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {Object}      returns the library name and version,
+   * @since 0.0.0
+   */
+  whoami() {
+    return this._library;
+  },
 
   /**
    * Returns a string.